consolidate logic for creating ingest pipelines
