+++ server/src/main/java/org/elasticsearch/action/ingest/PutPipelineTransportAction.java
@@ -59,42 +59,15 @@ public class PutPipelineTransportAction extends AcknowledgedTransportMasterNodeA
     @Override
     protected void masterOperation(Task task, PutPipelineRequest request, ClusterState state, ActionListener<AcknowledgedResponse> listener)
             throws Exception {
-
-        Map<String, Object> pipelineConfig = null;
-        IngestMetadata currentIngestMetadata = state.metadata().custom(IngestMetadata.TYPE);
-        if (currentIngestMetadata != null && currentIngestMetadata.getPipelines().containsKey(request.getId())) {
-            pipelineConfig = XContentHelper.convertToMap(request.getSource(), false, request.getXContentType()).v2();
-            var currentPipeline = currentIngestMetadata.getPipelines().get(request.getId());
-            if (currentPipeline.getConfigAsMap().equals(pipelineConfig)) {
-                // existing pipeline matches request pipeline -- no need to update
-                listener.onResponse(AcknowledgedResponse.TRUE);
-                return;
+        ingestService.putPipeline(
+            request,
+            listener,
+            (nodeListener) -> {
+                NodesInfoRequest nodesInfoRequest = new NodesInfoRequest();
+                nodesInfoRequest.clear();
+                nodesInfoRequest.addMetric(NodesInfoRequest.Metric.INGEST.metricName());
+                client.admin().cluster().nodesInfo(nodesInfoRequest, nodeListener);
             }
-        }
-
-        if (state.getNodes().getMinNodeVersion().before(Version.V_7_15_0)) {
-            pipelineConfig = pipelineConfig == null
-                ? XContentHelper.convertToMap(request.getSource(), false, request.getXContentType()).v2()
-                : pipelineConfig;
-            if (pipelineConfig.containsKey(Pipeline.META_KEY)) {
-                throw new IllegalStateException("pipelines with _meta field require minimum node version of " + Version.V_7_15_0);
-            }
-        }
-        NodesInfoRequest nodesInfoRequest = new NodesInfoRequest();
-        nodesInfoRequest.clear();
-        nodesInfoRequest.addMetric(NodesInfoRequest.Metric.INGEST.metricName());
-        client.admin().cluster().nodesInfo(
-            nodesInfoRequest,
-            ActionListener.wrap(
-                nodeInfos -> {
-                    Map<DiscoveryNode, IngestInfo> ingestInfos = new HashMap<>();
-                    for (NodeInfo nodeInfo : nodeInfos.getNodes()) {
-                        ingestInfos.put(nodeInfo.getNode(), nodeInfo.getInfo(IngestInfo.class));
-                    }
-                    ingestService.putPipeline(ingestInfos, request, listener);
-                },
-                listener::onFailure
-            )
         );
     }
